requirements.txt
flask
requests
beautifulsoup4
playwright
notion-client
python-dotenv
telegram
groq